Move board hook to the generic component
