+++ front/src/modules/companies/components/CompanyBoard/CompanyBoardContext.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const CompanyBoardContext = createContext<string | null>(null);
