--- front/src/pages/opportunities/BoardContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const BoardContext = createContext<string | null>(null);
