+++ front/src/pages/opportunities/Opportunities.tsx
@@ -1,25 +1,17 @@
-import { useCallback, useMemo } from 'react';
-import { getOperationName } from '@apollo/client/utilities';
 import { useTheme } from '@emotion/react';
 
-import { CompanyBoardCard } from '@/companies/components/CompanyBoard/CompanyBoardCard';
+import { CompanyBoardContext } from '@/companies/components/CompanyBoard/CompanyBoardContext';
 import { CompanyProgressBoard } from '@/companies/components/CompanyBoard/CompanyProgressBoard';
-import { NewCompanyProgressButton } from '@/companies/components/CompanyBoard/NewCompanyProgressButton';
 import { BoardActionBarButtonDeletePipelineProgress } from '@/pipeline-progress/components/BoardActionBarButtonDeletePipelineProgress';
 import { EntityBoardActionBar } from '@/pipeline-progress/components/EntityBoardActionBar';
 import { useBoard } from '@/pipeline-progress/hooks/useBoard';
-import { GET_PIPELINES } from '@/pipeline-progress/queries';
 import { RecoilScope } from '@/recoil-scope/components/RecoilScope';
 import { IconTargetArrow } from '@/ui/icons/index';
 import { WithTopBarContainer } from '@/ui/layout/containers/WithTopBarContainer';
 
 import {
-  PipelineProgress,
-  PipelineStage,
   useGetCompaniesQuery,
   useGetPipelinesQuery,
-  useUpdateOnePipelineProgressMutation,
-  useUpdateOnePipelineProgressStageMutation,
 } from '../../generated/graphql';
 
 export function Opportunities() {
@@ -33,69 +25,22 @@ export function Opportunities() {
     useGetCompaniesQuery,
   );
 
-  const columns = useMemo(
-    () =>
-      initialBoard?.map(({ id, colorCode, title }) => ({
-        id,
-        colorCode,
-        title,
-      })),
-    [initialBoard],
-  );
-  const [updatePipelineProgress] = useUpdateOnePipelineProgressMutation();
-  const [updatePipelineProgressStage] =
-    useUpdateOnePipelineProgressStageMutation();
-
-  const handleCardUpdate = useCallback(
-    async (
-      pipelineProgress: Pick<PipelineProgress, 'id' | 'amount' | 'closeDate'>,
-    ) => {
-      await updatePipelineProgress({
-        variables: {
-          id: pipelineProgress.id,
-          amount: pipelineProgress.amount,
-          closeDate: pipelineProgress.closeDate || null,
-        },
-        refetchQueries: [getOperationName(GET_PIPELINES) ?? ''],
-      });
-    },
-    [updatePipelineProgress],
-  );
-
-  const handleCardMove = useCallback(
-    async (
-      pipelineProgressId: NonNullable<PipelineProgress['id']>,
-      pipelineStageId: NonNullable<PipelineStage['id']>,
-    ) => {
-      updatePipelineProgressStage({
-        variables: {
-          id: pipelineProgressId,
-          pipelineStageId,
-        },
-      });
-    },
-    [updatePipelineProgressStage],
-  );
-
   return (
     <WithTopBarContainer
       title="Opportunities"
       icon={<IconTargetArrow size={theme.icon.size.md} />}
     >
       {items && pipelineId ? (
-        <>
+        <RecoilScope SpecificContext={CompanyBoardContext}>
           <CompanyProgressBoard
             pipelineId={pipelineId}
-            columns={columns || []}
             initialBoard={initialBoard}
             initialItems={items}
-            onCardMove={handleCardMove}
-            onCardUpdate={handleCardUpdate}
           />
           <EntityBoardActionBar>
             <BoardActionBarButtonDeletePipelineProgress />
           </EntityBoardActionBar>
-        </>
+        </RecoilScope>
       ) : (
         <></>
       )}
